+++ src/pages/index1.js
@@ -1,0 +1,14 @@
+export default function Home() {
+  
+  return (
+  <>
+    <head>
+      <title>HTML Meta Tag</title>
+      <meta httpEquiv = "refresh" url = "/home" />
+   </head>
+   <body>
+      <p>Redirecting to another URL</p>
+   </body>
+   </>
+   )
+}
